pyproject.toml
[tool.poetry]
name = "imaterialist"
version = "0.1.0"
description = ""
authors = ["hrsma2i <hrs.ma2i@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
kaggle = "^1.5.12"
Pillow = "^8.1.2"
segmentation = {git = "https://github.com/hrsma2i/segmentation.git", rev = "cfc1670021"}
typer = "^0.3.2"
pycocotools = "^2.0.4"

[tool.poetry.dev-dependencies]
numpy = "^1.20.1"
pandas = "^1.2.3"
matplotlib = "^3.3.4"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
